rename redux types file
